requirements.txt
# AI & Computer Vision
ultralytics==8.0.196
face-recognition==1.3.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.0.1

# Web Framework
Flask==3.0.0
Flask-Login==0.6.3
Werkzeug==3.0.1

# Telegram Bot
python-telegram-bot==20.7

# Database
sqlite3

# Utilities
requests==2.31.0
python-dotenv==1.0.0
